Adiciona soma com epsilon quando a linha sobrepoem um vértice do poligono

+++ TP01/Ponto.hpp
@@ -8,8 +8,8 @@ using namespace std;
 
 class Ponto {
 public:
-    long long int x, y;
-    Ponto(long long int x, long long int y);
+    double x, y;
+    Ponto(double x, double y);
 
     // Dado o segmento 'a-b', verifica se o ponto 'c' está sobre o segmento 'a-b'
     bool inSegment(Ponto a, Ponto b, Ponto c);
@@ -18,7 +18,7 @@ public:
     // 0 --> a, b e c são colineares
     // 1 --> sentido horário
     // 2 --> sentido anti-horário
-    int orientation(Ponto a, Ponto b, Ponto c);
+    double orientation(Ponto a, Ponto b, Ponto c);
 
     // Verifica se o segmento 'a-b' cruza com o segmento 'c-d'
     bool checkIntersect(Ponto a, Ponto b, Ponto c, Ponto d);
